requirements.txt
fastapi
uvicorn
google-generativeai
pdfplumber
fuzzywuzzy
python-Levenshtein
pydantic
python-multipart
python-dotenv
pillow
